selenium: fix upload of binary file to avoid sqlite deadlock
Looks like something in selenium has changed, but this code was dubious
to begin with. upload-id-upload is not a form but rather a (submit)
button. I suspect that the "submit" action was causing a click *and* a
submit at the same time, so two requests were getting dispatched,
causing a deadlock in sqlite.

The fix would either be to submit the form, or simply click on the
button. The latter is chosen to avoid any unnecessary changes.